test(proxy): split reverse parity coverage
